improved goal poses for the hello world
Singularities still cause issues with the current inverse kinematics and
it's possible, depending on the exact pose in the singularity for the second
move_joints command, that the triangle cannot be executed. So the poses are
slightly changed to avoid singularities and to ensure that the hello world
runs more reliable.

Signed-off-by: Wolfgang Wiedmeyer <wolfgang.wiedmeyer@kit.edu>