Small changes to the hello worlds
The move_joints call is repeated to ensure the right config for the
triangle movement. I removed some sleeps to make the movement sequence
more impressive.

The new IK does not handle the right corner of the triangle well because
there is an upcoming joint limit. The position is moved to the left in
accordance to the related rll_move test case.

Signed-off-by: Wolfgang Wiedmeyer <wolfgang.wiedmeyer@kit.edu>