Use the cap endeffector
Configured the project to use the new "cap" eef type instead of the gripper. The example code has been updated to reflect this change. The RViz config has been updated as well to show the correct cap TF. To better see what the project looks like a screenshot has been added to the Readme.