Tests: run playground.py in CI
Add a TestCase to run the playground.py via the CI. The lauch scripts have been adjusted to allow passing the use_sim parameter to specify whether or not to launch the script with Gazebo. The CI runs the script twice once with and once without gazebo. However, the gazebo-enabled version does seem to cause deviation errors from time to time.